+++ app/hotReload/hot.plugin.js
@@ -15,9 +15,9 @@ const throwError = (message) =>
 module.exports = {
   overrideWebpackConfig: ({ webpackConfig }) => {
     if (process.env.REACT_HOT_MODE === 'hot') {
-      const configTs = require('./components.tsconfig.json');
+      const configTs = require('./packages.tsconfig.json');
       if (!configTs) {
-        throwError('Please create components.tsconfig.json in hotReload repository');
+        throwError('Please create packages.tsconfig.json in hotReload repository');
       }
       console.log('adding ts-loader');
       const tsLoader = {
@@ -38,8 +38,30 @@ module.exports = {
         '@mui/styles': path.resolve(__dirname, '../../node_modules/@mui/styles'),
         '@mui/material': path.resolve(__dirname, '../../node_modules/@mui/material'),
         'react-i18next': path.resolve(__dirname, '../../node_modules/react-i18next'),
-        '@opentalk/common': path.resolve(__dirname, '../../node_modules/@opentalk/common'),
-        '@opentalk/components': path.resolve(__dirname, configTs.compilerOptions.paths['@opentalk/components'][0]),
+        '@opentalk/common': path.resolve(
+          __dirname,
+          configTs.compilerOptions.paths['@opentalk/common']
+            ? configTs.compilerOptions.paths['@opentalk/common'][0]
+            : '../../node_modules/@opentalk/common'
+        ),
+        '@opentalk/components': path.resolve(
+          __dirname,
+          configTs.compilerOptions.paths['@opentalk/components']
+            ? configTs.compilerOptions.paths['@opentalk/components'][0]
+            : '../node_modules/@opentalk/components'
+        ),
+        '@opentalk/rest-api-rtk-query': path.resolve(
+          __dirname,
+          configTs.compilerOptions.paths['@opentalk/rest-api-rtk-query']
+            ? configTs.compilerOptions.paths['@opentalk/rest-api-rtk-query'][0]
+            : '../../node_modules/@opentalk/rest-api-rtk-query'
+        ),
+        '@opentalk/react-redux-appauth': path.resolve(
+          __dirname,
+          configTs.compilerOptions.paths['@opentalk/react-redux-appauth']
+            ? configTs.compilerOptions.paths['@opentalk/react-redux-appauth'][0]
+            : '../../node_modules/@opentalk/react-redux-appauth'
+        ),
       };
       webpackConfig.resolve.alias = localDependenvies;