+++ packages/common/components/Notistack/fragments/index.ts
@@ -3,15 +3,7 @@
 // SPDX-License-Identifier: EUPL-1.2
 
 export { SnackbarProvider } from './SnackbarProvider';
-export {
-  ISnackActionsProps,
-  ISnackbarPersistentProps,
-  ISnackbarActionButtonProps,
-  notificationAction,
-  notificationPersistent,
-  notifications,
-} from './utils';
-export * from "./StackedMessages";
-export {
-  getNotistackComponents
-} from "./variations";
+export { notificationAction, notificationPersistent, notifications } from './utils';
+export type { ISnackActionsProps, ISnackbarPersistentProps, ISnackbarActionButtonProps } from './utils';
+export * from './StackedMessages';
+export { getNotistackComponents } from './variations';