+++ packages/common/index.ts
@@ -5,7 +5,6 @@
 export * from './components';
 export * from './assets/icons';
 export * from './types';
-export { ErrorStruct } from './utils';
 export * from './utils';
 export * from './store';
 export * from './hooks';